backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
sentence-transformers==3.0.0
pydantic==2.9.0
httpx==0.27.0
python-dotenv==1.0.0
numpy>=1.26.0
pystray>=0.19.0
Pillow>=10.0.0
pyinstaller>=6.0.0
